backend/requirements.txt
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
pypdf==5.3.1
striprtf==0.0.26
fastapi==0.115.11
uvicorn==0.34.0
pypinyin==0.55.0
thefuzz==0.22.1
httpx==0.27.0
